requirements.txt
about-time==4.2.1
absl-py==2.1.0
alabaster==0.7.13
albucore==0.0.9
albumentations==1.4.8
alive-progress==3.1.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
arabic-reshaper==3.0.0
asn1crypto==1.5.1
asttokens==2.4.1
attrs==23.2.0
autograd==1.6.2
Babel==2.15.0
backcall==0.2.0
backports.zoneinfo==0.2.1
boto3==1.34.122
botocore==1.34.122
build==1.2.1
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cma==3.2.2
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.1.1
coverage==5.3.1
cryptography==42.0.8
cssselect2==0.7.0
cycler==0.12.1
data-gradients==0.3.2
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
docutils==0.17.1
einops==0.3.2
executing==2.0.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
future==1.0.0
google-auth==2.30.0
google-auth-oauthlib==1.0.0
grapheme==0.6.0
grpcio==1.64.1
html5lib==1.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.7
imagededup==0.3.2
imageio==2.34.1
imagesize==1.4.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json-tricks==3.16.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jstyleson==0.0.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
nncf==2.11.0
numpy==1.23.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onemetric==0.1.2
onnx==1.15.0
onnxruntime==1.15.0
onnxsim==0.4.36
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openvino==2024.2.0
openvino-dev==2024.2.0
openvino-telemetry==2024.1.0
oscrypto==1.3.0
packaging==24.0
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip-tools==7.4.1
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pyDeprecate==0.3.2
pydot==2.0.0
Pygments==2.18.0
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
pymoo==0.6.1.1
pyparsing==3.1.2
pypdf==4.2.0
pypng==0.20220715.0
pyproject_hooks==1.1.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.0.3
qrcode==7.4.2
rapidfuzz==3.9.3
referencing==0.35.1
reportlab==3.6.13
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.0
seaborn==0.13.2
six==1.16.0
snowballstemmer==2.2.0
Sphinx==4.0.3
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.3
stringcase==1.2.0
super-gradients==3.7.1
supervision==0.21.0
svglib==1.5.1
sympy==1.12.1
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==1.1.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tinycss2==1.3.0
tomli==2.0.1
torch==2.3.1
torchmetrics==0.8.0
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
treelib==1.6.1
triton==2.3.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ultralytics==8.2.30
ultralytics-thop==0.2.8
uritools==4.0.3
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.16.0
xhtml2pdf==0.2.11
zipp==3.19.2
